// packages/files/src/gen/types/createFileUploadResponse.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Lemma Files API
 * OpenAPI spec version: 0.1.0
 */
import type { FileUpload } from "./fileUpload.js";
import type { FileUploadUrl } from "./fileUploadUrl.js";

export interface CreateFileUploadResponse {
  upload: FileUpload;
  uploadUrl: FileUploadUrl;
}