// packages/files/src/gen/types/file.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Lemma Files API
 * OpenAPI spec version: 0.1.0
 */
import type { FileContentType } from "./fileContentType.js";
import type { FilePurpose } from "./filePurpose.js";
import type { FileStatus } from "./fileStatus.js";

export interface File {
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  id: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  ownerUserId: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  createdByUserId: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  originalName: string;
  contentType: FileContentType;
  /**
   * @maximum 26214400
   * @exclusiveMinimum 0
   */
  byteSize: number;
  /** @pattern ^[A-Fa-f0-9]{64}$ */
  checksumSha256: string;
  status: FileStatus;
  purpose: FilePurpose;
  /** @nullable */
  deletedAt: string | null;
  createdAt: string;
  updatedAt: string;
}